# apps/api/docs/swagger.yaml
basePath: /
definitions:
  auth.LoginPayload:
    properties:
      email:
        type: string
      password:
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST para gestión de mascotas y usuarios.
  title: My Pets API
  version: "1.0"
paths:
...
  /api/v1/auth/google/callback:
    get:
      responses:
        "302":
          description: Redirect to frontend
        "400":
          description: invalid state or code
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Google OAuth callback
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/auth.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: 'data: User'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with email and password
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 'message: logged out'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: authentication required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Logout — clears auth cookies
      tags:
      - auth
  /api/v1/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 'data: User'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get the currently authenticated user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 'data: User'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: invalid or expired refresh token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token using the refresh cookie
      tags:
      - auth
...
securityDefinitions:
  CookieAuth:
    description: Cookie HttpOnly de sesión. Obtenida al hacer login o refresh. Se
      envía automáticamente por el navegador.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"
...